third_party/WebKit/LayoutTests/inspector/console/console-functions.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function onload()
{
    var functions = [
        function simple() {},
        async function asyncSimple() {},
        function* genSimple() {},
        function(){},
        function(x, y){},
        function namedArgs(x) {},
        function namedArgs2(x, y) {},
        function ({}) {},
        function   *    whitespace   (  x  )    {   },
        async    function    whitespace2   (  x  ,  y  ,  z  )    {   },
    ];

    var obj = {};
    for (var i = 0; i < functions.length; ++i) {
        console.log(functions[i]);
        console.dir(functions[i]);
        obj["func" + i] = functions[i];
    }
    console.log(obj);
    console.dir(obj);

    runTest();
}

function test() {
  ConsoleTestRunner.expandConsoleMessages(dumpConsoleMessages);

  function dumpConsoleMessages() {
    ConsoleTestRunner.dumpConsoleMessages(false, false, TestRunner.textContentWithLineBreaks);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="onload()">
<p>Tests that console logging different types of functions correctly.</p>
</body>
</html>
